refactor(gateway): split backend_client.rs (1491 lines) into focused submodules (#841)
The single 1491-line file is replaced by a backend_client/ directory
module containing six focused files:

  urls.rs     (35 lines)  - URL helper functions
  probe.rs   (144 lines)  - ProbeOutcome + probe_readiness/mcp_readiness
  error.rs    (69 lines)  - BackendCallError enum + Display
  http.rs    (168 lines)  - HTTP primitives (rest_get/post, post_jsonrpc,
                            parse_jsonrpc_result, percent_encode_uri,
                            uuid_like_id)
  ops.rs     (328 lines)  - All REST + JSON-RPC forward operations
                            (call_backend, fetch_*/try_fetch_*,
                            forward_tools/prompts_call,
                            subscribe_resource, read_resource)
  mod.rs      (49 lines)  - Module doc + pub(crate) mod declarations
                          + re-exports (all original public APIs preserved)
                          + inline #[cfg(test)] mod tests block

All existing public/pub(crate) symbols are re-exported from mod.rs so
no call-site changes are needed in gateway, aggregator, or http crates.

291 gateway tests pass unchanged.

Fixes #841
